requirements.txt
# Core dependencies
keyring>=24.0.0
requests>=2.31.0

# Optional: Gmail integration
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0

# Optional: Spotify integration
spotipy>=2.23.0

# Optional: Messaging
twilio>=8.10.0

# Optional: Voice
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.13